Fix exception when using multiple hubs
If you have multiple hubs with different methods on the hubs, an
exception would be thrown because a method could not be looked up. The
error is that instead of exiting the function to try the next hub, it would
try to execute the function on the hub that didn't have the method.